+++ doc/changelog.md
@@ -3,6 +3,7 @@
 ## 0.0.5 (Alpha)
 * Fixed Autostart
 * Hourly downloading of diagnosis keys (if available)
+* Improved syncing of diagnosis keys
 
 ## 0.0.4 (Alpha)
 * Prune collected tokens from database once a day